Client is self-contained in one file; no need for a source tree
