+++ src/ts/index.ts
@@ -3,7 +3,7 @@ import * as charts    from './charts/core'
 import * as factory   from './models/items/factory'
 import * as app       from './app/app'
 import * as edit      from './edit/edit'
-import { Client }     from './client'
+import Client         from './client'
 import { actions }    from './core/action'
 import { transforms } from './models/transform/transform'
 import User           from './models/user'
